[feat] 무한 스크롤링 구현

+++ src/public/post-read/post-read.js
@@ -1,7 +1,11 @@
 import { requestCancelLike, requestCreateLike } from '../api/post-like.js';
 import { requestReadPost, requestDeletePost } from '../api/posts.js';
 import { requestComments } from '../api/comments.js';
-import { paintCommentFormContainer, paintCommentsContainer } from '../component/comments/comments.js';
+import {
+    attachCommentEventHandler,
+    paintCommentFormContainer,
+    paintCommentsContainer,
+} from '../component/comments/comments.js';
 import { paintPostReadContainer } from '../component/post/post.js';
 import { paintHeader } from '../component/common/header/header.js';
 import { openModal } from '../component/common/modal/modal.js';
@@ -68,15 +72,45 @@ document.addEventListener('DOMContentLoaded', async () => {
     paintCommentsContainer(commentRes.data.comments, loginMemberId);
 
     /* 게시글 삭제 모달 */
-    document.querySelector('.delete-btn').addEventListener('click', ({ target }) => {
-        console.log(target);
-        const { domain, id } = target.dataset;
+    const postDeleteBtnElement = document.querySelector('.delete-btn');
 
-        openModal({
-            mainText: '게시글을 삭제하시겠습니까?',
-            subText: '삭제한 내용은 복구할 수 없습니다',
-            dataset: target.dataset,
-            onConfirm: () => deletePostHandler(id),
+    if (postDeleteBtnElement) {
+        postDeleteBtnElement.addEventListener('click', ({ target }) => {
+            console.log(target);
+            const { domain, id } = target.dataset;
+
+            openModal({
+                mainText: '게시글을 삭제하시겠습니까?',
+                subText: '삭제한 내용은 복구할 수 없습니다',
+                dataset: target.dataset,
+                onConfirm: () => deletePostHandler(id),
+            });
         });
+    }
+
+    /* 댓글 이벤트 등록 */
+    attachCommentEventHandler();
+
+    let isNewCommentFetching = false;
+
+    window.addEventListener('scroll', async () => {
+        const hitsBottom = window.innerHeight + window.scrollY >= document.documentElement.scrollHeight - 1;
+
+        if (!hitsBottom || isNewCommentFetching) {
+            return;
+        }
+
+        isNewCommentFetching = true;
+
+        const lastCommentId = Array.from(document.querySelectorAll('[data-commentid]')).at(-1).dataset.commentid;
+        const res = await requestComments(postId, { lastCommentId });
+
+        if (!res.success) {
+            document.querySelector('.comments-container').textContent = res.data;
+            return;
+        }
+
+        paintCommentsContainer(res.data.comments, loginMemberId);
+        isNewCommentFetching = false;
     });
 });
